changed deleteCourseCategoryById to boolean in ServiceImpl
also moved business logic from controller to service

+++ src/main/java/com/fujitsu/ph/tsup/course/category/service/CourseCategoryManagementServiceImpl.java
@@ -69,8 +69,14 @@ public class CourseCategoryManagementServiceImpl implements CourseCategoryManage
     }
 
     @Override
-    public void deleteCourseCategoryById(Long id) {
-        courseCategoryManagementDao.deleteCourseCategoryById(id);
+    public boolean deleteCourseCategoryById(Long id) {
+        CourseCategory courseCategoryResult = courseCategoryManagementDao.findCourseCategoryById(id);
+        if (courseCategoryResult != null) {
+            courseCategoryManagementDao.deleteCourseCategoryById(id);
+            return true;
+        } else {
+            return false;
+        }        
     }
 
     @Override
